+++ system-parachains/asset-hubs/asset-hub-polkadot/src/ah_migration/call_filter.rs
@@ -126,11 +126,10 @@ pub fn call_allowed_status(
 		}) => ON, // Only root
 		SnowbridgeSystemFrontend(..) => OFF,
 		Staking(..) => OFF,
-		StakingRcClient(..) => ON,     // Keep on for incoming RC calls over XCM
-		StateTrieMigration(..) => OFF, // Deprecated
-		System(..) => ON,              // remark plus root calls
-		Timestamp(..) => ON,           // only `set` inherit
-		ToKusamaXcmRouter(..) => ON,   // Allow to report bridge congestion
+		StakingRcClient(..) => ON,   // Keep on for incoming RC calls over XCM
+		System(..) => ON,            // remark plus root calls
+		Timestamp(..) => ON,         // only `set` inherit
+		ToKusamaXcmRouter(..) => ON, // Allow to report bridge congestion
 		Treasury(..) => OFF,
 		Uniques(..) => OFF,
 		Utility(..) => ON, // batching etc, just convenience
@@ -201,7 +200,6 @@ pub fn call_allowed_before_migration(
 		Session(..) |
 		SnowbridgeSystemFrontend(..) |
 		StakingRcClient(..) |
-		StateTrieMigration(..) |
 		System(..) |
 		Timestamp(..) |
 		ToKusamaXcmRouter(..) |